使用泛型接口  链表实现栈和双端队列
使用泛型接口  链表实现栈和双端队列

+++ src/test/FirstLastLinkTest.java
@@ -8,12 +8,12 @@ public class FirstLastLinkTest {
 	
 	public static void run(){
 		theLink=new FirstLastLink();
-		theLink.insertFirst(22);
-		theLink.insertFirst(44);
-		theLink.insertFirst(66);
-		theLink.insertLast(11);
-		theLink.insertLast(33);
-		theLink.insertLast(55);
+		theLink.insertFirst((long) 22);
+		theLink.insertFirst((long) 44);
+		theLink.insertFirst((long) 66);
+		theLink.insertLast((long) 11);
+		theLink.insertLast((long) 33);
+		theLink.insertLast((long) 55);
 		theLink.displayList();
 		theLink.deleteFirst();
 		theLink.deleteFirst();
